RealTimeDataProcessing/requirements.txt
kafka-python==2.0.2
Faker==19.2.0
pyspark==3.1.2
requests